% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcglm_methods.R
\name{coef.mcmcglm}
\alias{coef.mcmcglm}
\title{S3 method for getting the average value of coefficients}
\usage{
\method{coef}{mcmcglm}(x)
}
\arguments{
\item{x}{an \code{mcmcglm} object}
}
\description{
S3 method for getting the average value of coefficients
}
