% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_performance.R
\name{compare_eta_comptime_across_nvars}
\alias{compare_eta_comptime_across_nvars}
\title{Compare runtime using CGGibbs and naive approach to calculate linear predictor}
\usage{
compare_eta_comptime_across_nvars(
  n_vars,
  n = 100,
  beta_prior = distributional::dist_normal(0, 1),
  log_likelihood_extra_args = list(sd = 1),
  sample_method = c("slice_sampling", "normal-normal"),
  qslice_fun = qslice::slice_stepping_out,
  ...,
  n_samples = 500,
  burnin = 100,
  parallelise = FALSE,
  n_cores = as.numeric(Sys.getenv("NUMBER_OF_PROCESSORS")) - 1
)
}
\arguments{
\item{n_vars}{a \verb{numeric vector} with numbers of normally distributed columns in the design matrix
for different runs}

\item{n}{a \code{numeric} with the number of observations in the data analysed by the GLM}

\item{beta_prior}{a \code{distribution} object created by a function from the
\code{\link{distributional}} package. Could fx. be \code{distributional::dist_normal(mean = 0, sd = 1)}.}

\item{log_likelihood_extra_args}{a named \code{list} with arguments passed onto the \link{log_density} function.
Fx. specification of \code{log_likelihood_extra_args = list(sd = x)} is needed for the case of
\code{family = "gaussian"}}

\item{sample_method}{a \code{character} specifying the method used for sampling. The default \code{"slice_sampling"}
is the intended value in most cases, as it works for any specification of \code{family} and \code{beta_prior}.
\code{"normal-normal"} uses a conditional normal distribution to sample from in case of conjugate prior with
gaussian response and \code{beta_prior}. Implemented for testing purposes but works for that niche case.}

\item{qslice_fun}{a \code{function} from the
\href{https://CRAN.R-project.org/package=qslice}{qslice} package.
Default is \link[qslice:slice_stepping_out]{qslice::slice_stepping_out} which uses the slice sampler from
\href{https://projecteuclid.org/journals/annals-of-statistics/volume-31/issue-3/Slice-sampling/10.1214/aos/1056562461.full}{Neal 2003},
but all functions are available.}

\item{...}{arguments passed onto the function specified by \code{qslice_fun}. For default \link[qslice:slice_stepping_out]{qslice::slice_stepping_out}
\code{w} needs to be specified, while for fx. \link[qslice:slice_elliptical]{qslice::slice_elliptical}, \code{mu} and \code{sigma} need to be specified}

\item{n_samples}{a \code{numeric} with number of samples to draw of each parameter(/variable) in the model}

\item{burnin}{a \code{numeric} with the number of samples to be marked as "burnin". Burnin samples are not
included in the \code{beta_mean} calculation to increase finite sample performance of the LLN estimate}

\item{parallelise}{a \code{logical} if the runs of the algorithm across values of \code{n_vars} should be
parallelised using \link[future.apply:future_lapply]{future.apply::future_lapply}}

\item{n_cores}{a \code{numeric} with the number of cores to use for parallelisation. Default is 1 less
than the number of available cores}
}
\value{
a \code{data.frame} with information on computation time for different values of \code{n_vars}
}
\description{
The comparison of the methods is described in
\href{https://mathiaslj.github.io/mcmcglm/articles/pospkg.html#advantage-of-cggibbs}{this vignette}.
The user can specify different arguments to alter the specification of the comparison, but it's
possible to only specify values of the \code{n_vars} argument. It's not possible to change the fact that
data is simulated with a normal response and normal explanatory variables.
}
\examples{
# Compare the runtime for 2, 20, 40, 60 variables in the model
compare_eta_comptime_across_nvars(n_vars = c(2, seq(from = 20, to = 60, by = 20)),
                                  n_samples = 100,
                                  burnin = 20)
}
