% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_utils.R
\name{log_density}
\alias{log_density}
\title{S3 generic for calculating the log density of a distribution dispatched via a \link{family}}
\usage{
log_density(family, mu, Y, ...)
}
\arguments{
\item{family}{a description of the error distribution and link function to be used in the model.
This can be a \code{character} string naming a family function, a family function or the result
of a call to a family function. (See \link{family} for details of family functions.)}

\item{mu}{a \verb{numeric vector} with values of the "main" parameter of the distribution
specified by the \code{family} argument}

\item{Y}{a \verb{numeric vector} of the response variable in which to evaluate the density}

\item{...}{arguments passed on to relevant methods}
}
\value{
\verb{numeric vector} of log_density values
}
\description{
The methods are parametrised by "mu" and takes additional arguments that are
needed for the calculation through the \code{...} argument
}
\details{
Supported families are \code{gaussian}, \code{binomial}, \code{poisson} and \code{negative.binomial}.
Implement your own S3 method to add support for a new family.

In \link{mcmcglm} when this function is called, the \code{mu} is the modelled
mean in the glm model (meaning it's the inverse link of the linear predictor). Reference
methods to see the parametrisation
}
