% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_utils.R
\name{log_potential_from_betaj}
\alias{log_potential_from_betaj}
\title{Calculate the log-potential (log-likelihood plus log-density of prior)}
\usage{
log_potential_from_betaj(
  new_beta_j,
  j,
  current_beta,
  current_eta,
  Y,
  X_j,
  family,
  beta_prior,
  ...
)
}
\arguments{
\item{new_beta_j}{a \code{numeric} new value of the j'th component of the beta parameter vector}

\item{j}{a \code{numeric} with the index of the parameter vector}

\item{current_beta}{the current value of the beta parameter vector in the sampling procedure}

\item{current_eta}{the current value of the linear predictor corresponding to the \code{current_beta} value}

\item{Y}{a \verb{numeric vector} of the response variable in which to evaluate the density}

\item{X_j}{the j'th column in the design matrix}

\item{family}{a description of the error distribution and link function to be used in the model.
This can be a \code{character} string naming a family function, a family function or the result
of a call to a family function. (See \link{family} for details of family functions.)}

\item{beta_prior}{a \code{distribution} object created by a function from the
\code{\link{distributional}} package. Could fx. be \code{distributional::dist_normal(mean = 0, sd = 1)}.}

\item{...}{arguments passed on to the S3 generic \link{log_density}}
}
\value{
The value of the log-potential having changed the j'th component of the \code{current_beta}
to \code{new_beta_j}
}
\description{
Calculates the log-potential as a function of a new coordinate of the beta parameter
vector. Done like this to use the unexporte
}
