% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_utilities.R
\name{plot_mcmcglm_across_tuningparams}
\alias{plot_mcmcglm_across_tuningparams}
\title{Plot a list of mcmcglms showing varying tuning parameters in the title of the plots}
\usage{
plot_mcmcglm_across_tuningparams(list_mcmcglms)
}
\arguments{
\item{list_mcmcglms}{A \code{list} of \code{mcmcglm} objects. Intended to be the result of
a call to \link{mcmcglm_across_tuningparams}}
}
\description{
Plot a list of mcmcglms showing varying tuning parameters in the title of the plots
}
\examples{
Create test data for different scenarios
n <- 100
x1 <- rnorm (n)
x2 <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2
lin_pred <- b0+b1*x1+b2*x2

#############################################
# Different families and priors

# For family "gaussian" and iid normal prior
y_norm <- rnorm(n, mean = lin_pred, sd = 1)
dat_norm <- data.frame(Y = y_norm, X1 = x1, X2 = x2)

w05_mcmcglms <- mcmcglm_across_tuningparams(
   seq(from = 0.5, by = 0.5, length.out = 9),
   tuning_parameter_name = "w",
   formula = Y ~ .,
   family = "gaussian",
   data = dat_norm,
   n_samples = 10,
   burnin = 0
)

plot_mcmcglm_across_tuningparams(w05_mcmcglms)
}
