% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcglm_methods.R
\name{quantile.mcmcglm}
\alias{quantile.mcmcglm}
\title{S3 method for getting quantiles of samples for each parameter}
\usage{
\method{quantile}{mcmcglm}(x, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{x}{an \code{mcmcglm} object}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The function summarises only samples that are not marked as burnin in samples(.)
}
\examples{
\dontrun{
# Create test data
n <- 100
x1 <- rnorm (n)
x2 <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2
lin_pred <- b0+b1*x1+b2*x2

# For family "gaussian" and iid normal prior
y_norm <- rnorm(n, mean = lin_pred, sd = 1)
dat_norm <- data.frame(Y = y_norm, X1 = x1, X2 = x2)

norm <- mcmcglm(formula = Y ~ .,
                   data = dat_norm,
                   beta_prior = distributional::dist_normal(0, 1),
                   family = "gaussian",
                   n_samples = 100,
                   burnin = 10,
                   sample_method = "slice_sampling",
                   qslice_fun = qslice::slice_stepping_out,
                   w = 0.5)
quantile(norm)
}


}
