% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcglm_methods.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a trace plot of the MCMC samples}
\usage{
trace_plot(x, samples_drop = NULL)
}
\arguments{
\item{x}{an \code{mcmcglm} object}

\item{samples_drop}{a \code{numeric} specifying a number of initial samples to
exclude from the trace_plot to improve the axis zoom on the plot}
}
\description{
Create a trace plot of the MCMC samples
}
