% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_utils.R
\name{update_linear_predictor}
\alias{update_linear_predictor}
\title{Update value of a linear predictor as function of a single coordinate change}
\usage{
update_linear_predictor(new_beta_j, current_beta_j, current_eta, X_j)
}
\arguments{
\item{new_beta_j}{a \code{numeric} new value of the j'th component of the beta parameter vector}

\item{current_beta_j}{the \code{numeric} current value of the j'th component of the beta parameter vector}

\item{current_eta}{the current value of the linear predictor corresponding to the \code{current_beta} value}

\item{X_j}{the j'th column in the design matrix}
}
\description{
Function for updating the linear predictor with n actions rather than n\emph{n_vars actions
if naively doing a matrix-vector product of X \%}\% beta
}
