% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcglm-package.R
\docType{package}
\name{mcmcglm-package}
\alias{mcmcglm-package}
\title{mcmcglm: Gibbs Sampling Methods}
\description{
Sample posterior distributions using variations of Gibbs samplers
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mathiaslj/mcmcglm}
  \item Report bugs at \url{https://github.com/mathiaslj/mcmcglm/issues}
}

}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

}
\keyword{internal}
